fixed elevator code (encoder distance calculations + hysteresis modified) - compile to 2019_DeepSpace!

+++ code/2019/include/ElevatorMech.h
@@ -42,7 +42,7 @@ public:
         void StartUpInit();
         //void FindLocation(double amountToMove);
 
-        int GetEncoderCount(); 
+        double GetEncoderCount(); 
         double GetEncoderDistance(); 
 
         void ElevatorMotorUp();
